changed login page from layout to page
